+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/temp/PlannerRuleSet.java
@@ -42,6 +42,7 @@ import com.apple.foundationdb.record.query.plan.temp.rules.InComparisonToExplode
 import com.apple.foundationdb.record.query.plan.temp.rules.MatchIntermediateRule;
 import com.apple.foundationdb.record.query.plan.temp.rules.MatchLeafRule;
 import com.apple.foundationdb.record.query.plan.temp.rules.MergeFetchIntoCoveringIndexRule;
+import com.apple.foundationdb.record.query.plan.temp.rules.MergeProjectionAndFetchRule;
 import com.apple.foundationdb.record.query.plan.temp.rules.NormalizePredicatesRule;
 import com.apple.foundationdb.record.query.plan.temp.rules.OrToLogicalUnionRule;
 import com.apple.foundationdb.record.query.plan.temp.rules.PushDistinctBelowFilterRule;
@@ -76,6 +77,7 @@ import java.util.stream.Stream;
  * A set of rules for use by a planner that supports quickly finding rules that could match a given planner expression.
  */
 @API(API.Status.EXPERIMENTAL)
+@SuppressWarnings("java:S1452")
 public class PlannerRuleSet {
     private static final List<PlannerRule<? extends RelationalExpression>> NORMALIZATION_RULES = ImmutableList.of(
             new NormalizePredicatesRule()
@@ -111,6 +113,7 @@ public class PlannerRuleSet {
             new PushSetOperationThroughFetchRule<>(RecordQueryUnorderedUnionPlan.class),
             new PushSetOperationThroughFetchRule<>(RecordQueryInUnionPlan.class),
             new RemoveProjectionRule(),
+            new MergeProjectionAndFetchRule(),
             new PushReferencedFieldsThroughDistinctRule(),
             new PushReferencedFieldsThroughFilterRule(),
             new PushReferencedFieldsThroughSelectRule(),
@@ -132,13 +135,13 @@ public class PlannerRuleSet {
             new DataAccessRule(),
             new SelectDataAccessRule()
     );
-    private static final List<PlannerRule<? extends RelationalExpression>> ALL_RULES =
+    private static final List<PlannerRule<? extends RelationalExpression>> ALL_EXPRESSION_RULES =
             ImmutableList.<PlannerRule<? extends RelationalExpression>>builder()
                     .addAll(EXPLORATION_RULES)
                     .addAll(IMPLEMENTATION_RULES)
                     .build();
 
-    public static final PlannerRuleSet ALL = new PlannerRuleSet(ALL_RULES);
+    public static final PlannerRuleSet DEFAULT = new PlannerRuleSet(ALL_EXPRESSION_RULES);
 
     @Nonnull
     private final Multimap<Class<?>, PlannerRule<? extends RelationalExpression>> ruleIndex =
@@ -166,7 +169,7 @@ public class PlannerRuleSet {
     @Nonnull
     public Stream<PlannerRule<? extends RelationalExpression>> getExpressionRules(@Nonnull RelationalExpression expression,
                                                                                   @Nonnull final Predicate<PlannerRule<? extends RelationalExpression>> rulePredicate) {
-        return Streams.concat(ruleIndex.get(expression.getClass()).stream().filter(rulePredicate), alwaysRules.stream());
+        return Streams.concat(ruleIndex.get(expression.getClass()).stream(), alwaysRules.stream()).filter(rulePredicate);
     }
 
     @Nonnull
@@ -176,7 +179,8 @@ public class PlannerRuleSet {
 
     @Nonnull
     public Stream<PlannerRule<? extends PartialMatch>> getPartialMatchRules(@Nonnull final Predicate<PlannerRule<? extends PartialMatch>> rulePredicate) {
-        return PARTIAL_MATCH_RULES.stream();
+        return PARTIAL_MATCH_RULES.stream()
+                .filter(rulePredicate);
     }
 
     @Nonnull
@@ -186,6 +190,12 @@ public class PlannerRuleSet {
 
     @Nonnull
     public Stream<PlannerRule<? extends MatchPartition>> getMatchPartitionRules(@Nonnull final Predicate<PlannerRule<? extends MatchPartition>> rulePredicate) {
-        return MATCH_PARTITION_RULES.stream();
+        return MATCH_PARTITION_RULES.stream()
+                .filter(rulePredicate);
+    }
+
+    @Nonnull
+    public Stream<? extends PlannerRule<?>> getAllRules() {
+        return Streams.concat(ALL_EXPRESSION_RULES.stream(), PARTIAL_MATCH_RULES.stream(), MATCH_PARTITION_RULES.stream());
     }
 }